+++ Tools/dump_platformlist.sh
@@ -25,7 +25,7 @@ if [[ ! -s /tmp/org.rehabman.platformlist.plist ]]; then
     exit
 fi
 
-for x in preinit native patched; do
+for x in native patched; do
     extract_bin_property $plist :0:platform-table-$x $x.bin
     xxd -groupsize 4 <$x.bin >$x.txt
 done