+++ WhateverGreen/kern_igfx.cpp
@@ -149,9 +149,10 @@ void IGFX::processKernel(KernelPatcher &patcher, DeviceInfo *info) {
 #ifdef DEBUG
 		if (checkKernelArgument("-igfxdump"))
 			dumpFramebufferToDisk = true;
+
+		if (checkKernelArgument("-igfxfbdump"))
+			dumpPlatformTable = true;
 #endif
-		if (checkKernelArgument("-igfxnoigdump"))
-			dumpPlatformTable = false;
 
 		bool connectorLessFrame = info->reportedFramebufferIsConnectorLess;
 
@@ -235,20 +236,17 @@ bool IGFX::processKext(KernelPatcher &patcher, size_t index, mach_vm_address_t a
 		}
 
 		if (applyFramebufferPatch || dumpFramebufferToDisk || hdmiAutopatch) {
-			if (cpuGeneration != CPUInfo::CpuGeneration::SandyBridge) {
-				gPlatformInformationList = patcher.solveSymbol<void *>(index, "_gPlatformInformationList", address, size);
-				gPlatformListIsSNB = false;
-			} else {
-				gPlatformInformationList = patcher.solveSymbol<void *>(index, "_PlatformInformationList", address, size);
+			if (cpuGeneration == CPUInfo::CpuGeneration::SandyBridge) {
 				gPlatformListIsSNB = true;
+				gPlatformInformationList = patcher.solveSymbol<void *>(index, "_PlatformInformationList", address, size);
+			} else {
+				gPlatformListIsSNB = false;
+				gPlatformInformationList = patcher.solveSymbol<void *>(index, "_gPlatformInformationList", address, size);
 			}
-			gPlatformInformationList = patcher.solveSymbol<void *>(index, cpuGeneration != CPUInfo::CpuGeneration::SandyBridge ? "_gPlatformInformationList" : "_PlatformInformationList", address, size);
+
 			if (gPlatformInformationList) {
 				framebufferStart = reinterpret_cast<uint8_t *>(address);
 				framebufferSize = size;
-
-				if (callbackIGFX->dumpPlatformTable)
-					callbackIGFX->writePlatformListData("platform-table-preinit");
 
 				auto fbGetOSInformation = "__ZN31AppleIntelFramebufferController16getOSInformationEv";
 				if (cpuGeneration == CPUInfo::CpuGeneration::SandyBridge)
@@ -359,9 +357,6 @@ bool IGFX::wrapAcceleratorStart(IOService *that, IOService *provider) {
 }
 
 uint64_t IGFX::wrapGetOSInformation(void *that) {
-	// call original first
-	uint64_t r = FunctionCast(wrapGetOSInformation, callbackIGFX->orgGetOSInformation)(that);
-
 #ifdef DEBUG
 	if (callbackIGFX->dumpFramebufferToDisk) {
 		char name[64];
@@ -371,18 +366,22 @@ uint64_t IGFX::wrapGetOSInformation(void *that) {
 	}
 #endif
 
+#ifdef DEBUG
 	if (callbackIGFX->dumpPlatformTable)
 		callbackIGFX->writePlatformListData("platform-table-native");
+#endif
 
 	if (callbackIGFX->applyFramebufferPatch)
 		callbackIGFX->applyFramebufferPatches();
 	else if (callbackIGFX->hdmiAutopatch)
 		callbackIGFX->applyHdmiAutopatch();
 
+#ifdef DEBUG
 	if (callbackIGFX->dumpPlatformTable)
 		callbackIGFX->writePlatformListData("platform-table-patched");
+#endif
 
-	return r;
+	return FunctionCast(wrapGetOSInformation, callbackIGFX->orgGetOSInformation)(that);
 }
 
 bool IGFX::wrapLoadGuCBinary(void *that, bool flag) {
@@ -711,12 +710,17 @@ uint8_t *IGFX::findFramebufferId(uint32_t framebufferId, uint8_t *startingAddres
 	return nullptr;
 }
 
+#ifdef DEBUG
 size_t IGFX::calculatePlatformListSize(size_t maxSize) {
+	// sanity check maxSize
+	if (maxSize < sizeof(uint32_t)*2)
+		return maxSize;
 	// ig-platform-id table ends with 0xFFFFF, but to avoid false positive
 	// look for FFFFFFFF 00000000
+	// and Sandy Bridge is special, ending in 00000000 000c0c0c
 	uint8_t * startingAddress = reinterpret_cast<uint8_t *>(gPlatformInformationList);
 	uint32_t *startAddress = reinterpret_cast<uint32_t *>(startingAddress);
-	uint32_t *endAddress = reinterpret_cast<uint32_t *>(startingAddress + maxSize);
+	uint32_t *endAddress = reinterpret_cast<uint32_t *>(startingAddress + maxSize - sizeof(uint32_t));
 	while (startAddress < endAddress) {
 		if ((!gPlatformListIsSNB && 0xffffffff == startAddress[0] && 0 == startAddress[1]) ||
 			(gPlatformListIsSNB && 0 == startAddress[0] && 0x0c0c0c00 == startAddress[1]))
@@ -728,10 +732,17 @@ size_t IGFX::calculatePlatformListSize(size_t maxSize) {
 }
 
 void IGFX::writePlatformListData(const char *subKeyName) {
-	auto wegEntry = IORegistryEntry::fromPath("IOService:/IOResources/WhateverGreen");
-	auto tableData = OSData::withBytes(gPlatformInformationList, (unsigned)calculatePlatformListSize(PAGE_SIZE));
-	wegEntry->setProperty(subKeyName, tableData);
+	auto entry = IORegistryEntry::fromPath("IOService:/IOResources/WhateverGreen");
+	if (entry) {
+		auto table = OSData::withBytes(gPlatformInformationList, static_cast<unsigned>(calculatePlatformListSize(PAGE_SIZE)));
+		if (table) {
+			entry->setProperty(subKeyName, table);
+			table->release();
+		}
+		entry->release();
+	}
 }
+#endif
 
 bool IGFX::applyPatch(const KernelPatcher::LookupPatch &patch, uint8_t *startingAddress, size_t maxSize) {
 	bool r = false;